PostgreSQL/requirements.txt
django
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
drf-spectacular
